MigrationTool/requirements.txt
google-cloud-firestore>=2.0.0
google-cloud-storage>=2.0.0
google-auth>=2.0.0
cryptography>=41.0.0
